Refactored and reorganized static classes into categorized Utility classes
