+++ src/main/java/TransformUtils.java
@@ -1,7 +1,7 @@
 import java.util.Arrays;
 import java.util.Random;
 
-public class BytePartitioner {
+public class TransformUtils {
     public static byte[][] splitWithRemainder (byte[] fileBytes, int n) throws Exception {
         int remainder = fileBytes.length % n;
         int fragSize = ((fileBytes.length - (remainder)) / n);
